add zn_cursor_move_relative
